+++ packages/react-static/src/static/getConfig.js
@@ -124,20 +124,25 @@ export function buildConfig(state, config = {}) {
   let assetsPath = ''
   if (process.env.REACT_STATIC_ENV === 'development') {
     basePath = cleanSlashes(config.devBasePath)
-    assetsPath = cleanSlashes(config.devAssetsPath || paths.assets)
+    assetsPath = config.devAssetsPath || paths.assets
   } else if (state.staging) {
     siteRoot = cutPathToRoot(config.stagingSiteRoot || '/', '$1')
     basePath = cleanSlashes(config.stagingBasePath)
-    assetsPath = cleanSlashes(config.stagingAssetsPath || paths.assets)
+    assetsPath = config.stagingAssetsPath || paths.assets
   } else {
     siteRoot = cutPathToRoot(config.siteRoot, '$1')
     basePath = cleanSlashes(config.basePath)
-    assetsPath = cleanSlashes(config.assetsPath || paths.assets)
+    assetsPath = config.assetsPath || paths.assets
   }
   const publicPath = `${cleanSlashes(`${siteRoot}/${basePath}`)}/`
- 
+
   if (assetsPath && !isAbsoluteUrl(assetsPath)) {
     assetsPath = `/${cleanSlashes(`${basePath}/${assetsPath}`)}/`
+  }
+
+  // add trailing slash only if assetsPath was supplied, but no trailing slash
+  if (assetsPath !== '' && !assetsPath.endsWith('/')) {
+    assetsPath = `${assetsPath}/`
   }
 
   // Add the project root as a plugin. This allows the dev